requirements.txt
pandas
tensorflow
scikit-learn
matplotlib
pillow
numpy